Moved graphics functions to own module, allowed loading from save
Functions and variables for gameGraphics.py were moved to graphicsFunctions.py
and both the option to load from a save file and the option to save to the
save file (when a new game is started, for example) are now available.
The MyAdventure() class is used to initialize variables relating to the
pyGame font to be used, screen and screen dimensions, as well as the
initial user position on the screen.